requirements.txt
spacy>=3.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pyspellchecker>=0.8.0
lemminflect>=0.2.3
